jira/requirements.txt
pygithub
pyyaml
jira
